move filesto components folder
